requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
python-vlc==3.0.21203
numpy==1.26.4
